Avoid abbreviations on the title, reformat
Until now, CLI was in the title. But that is discouraged. We'd better avoid
abbreviations as much as we can.

Also, it is better to put sentences on one line for better version
controlling. This way, when we change a sentence or a word, it can be
distinguished easily.

This commit fixed the problems above.